backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.10
pytest==7.4.3
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
asyncpg==0.29.0
bcrypt==4.1.2
pyjwt==2.8.0
python-multipart==0.0.6
email-validator==2.1.0
